+++ tests/resources/user_role_test.go
@@ -1,0 +1,26 @@
+package tests
+
+import (
+	"testing"
+
+	"github.com/hashicorp/terraform-plugin-testing/config"
+	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+)
+
+func TestUserRole(t *testing.T) {
+	t.Run("CRUD", func(t *testing.T) {
+		resource.Test(t, resource.TestCase{
+			ProtoV6ProviderFactories: providerFactory,
+			Steps: []resource.TestStep{
+				{
+					Config:          providerConfigUs,
+					ConfigDirectory: config.TestNameDirectory(),
+					Check: resource.ComposeAggregateTestCheckFunc(
+						resource.TestCheckResourceAttrSet("konnect_user_role.my_userrole", "entity_id"),
+						resource.TestCheckResourceAttr("konnect_user_role.my_userrole", "role_name", "Viewer"),
+					),
+				},
+			},
+		})
+	})
+}